+++ src/components/Editor.svelte
@@ -1,6 +1,5 @@
 <script>
   import { onMount } from "svelte";
-  import * as StrongsConcordance from '../lib/concordances/strongs/finder';
   import Box from "./Box.svelte";
   import NounEditor from "./NounEditor.svelte";
   import { GlobalEventBus } from "../lib/events";
@@ -43,9 +42,6 @@
     quill.focus();
     quill.on('text-change', () => words = getText())
   });
-
-  // example of finder code
-  console.log(StrongsConcordance.findReplacements_v3('David'))
 
   function getText() {
     if (!quill) return []