+++ src/lib/ety-js/data/index.js
@@ -1,0 +1,9 @@
+import etymologies from './etymologies.json'
+import iso_639_2 from './iso-639-2.json'
+import iso_639_3 from './iso-639-3.json'
+
+export default {
+  etymologies,
+  iso_639_2,
+  iso_639_3,
+}